+++ ricaun.NUnit/Extensions/TaskSTA.cs
@@ -1,0 +1,62 @@
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace ricaun.NUnit.Extensions
+{
+    /// <summary>
+    /// TaskSTA
+    /// </summary>
+    internal static class TaskSTA
+    {
+        /// <summary>
+        /// Run
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <param name="func"></param>
+        /// <returns></returns>
+        public static Task<T> Run<T>(Func<T> func)
+        {
+            var tcs = new TaskCompletionSource<T>();
+            Thread thread = new Thread(() =>
+            {
+                try
+                {
+                    tcs.SetResult(func());
+                }
+                catch (Exception e)
+                {
+                    tcs.SetException(e);
+                }
+            });
+            thread.SetApartmentState(ApartmentState.STA);
+            thread.Start();
+            return tcs.Task;
+        }
+
+        /// <summary>
+        /// Run
+        /// </summary>
+        /// <param name="action"></param>
+        /// <returns></returns>
+        public static Task Run(Action action)
+        {
+            var tcs = new TaskCompletionSource<object>();
+            Thread thread = new Thread(() =>
+            {
+                try
+                {
+                    action();
+                    tcs.SetResult(null);
+                }
+                catch (Exception e)
+                {
+                    tcs.SetException(e);
+                }
+            });
+            thread.SetApartmentState(ApartmentState.STA);
+            thread.Start();
+            return tcs.Task;
+        }
+    }
+}